Safari-611.1.10/LayoutTests/css3/flexbox/auto-height-dynamic.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.flexbox {
    width: 600px;
    background-color: #aaa;
    position: relative;
}
.flexbox :nth-child(1) {
    background-color: blue;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

<div class="flexbox column" data-expected-height=0>
    <div class="flexitem" data-expected-height=0 style="flex: 100px"></div>
</div>

<div class="flexbox column" data-expected-height=0>
    <div class="flexitem2" data-expected-height=0 style="flex: 100px"></div>
</div>

<script>
document.querySelector('.flexitem').offsetHeight;
document.querySelector('.flexitem2').offsetHeight;
document.querySelector('.flexitem').style.flex = "1";
document.querySelector('.flexitem2').style.flex = "1 auto";
</script>

</body>
</html>
